+++ Assets/Scripts/GetOS.cs
@@ -1,4 +1,4 @@
-﻿using UnityEngine;
+using UnityEngine;
 using System.Collections;
 using System.Diagnostics;
 using UnityEngine.UI;
@@ -16,6 +16,7 @@ public class GetOS : MonoBehaviour
     {
         string platform = null;
         int id = 0;
+        title.text = "Scammer Bingo Unix";
         if (!Application.isEditor)
         {
 
@@ -47,6 +48,11 @@ public class GetOS : MonoBehaviour
             title.text = "Scammer Bingo: OSX Edition";
             platform = "osx";
         }
+        if (Application.platform == RuntimePlatform.WindowsPlayer) 
+        {
+            title.text = "Scammer Bingo: Windows Test Environment";
+            platform = "windows";
+        }
         if (Application.platform == RuntimePlatform.WindowsEditor || Application.platform == RuntimePlatform.OSXEditor)
         {
             title.text = "Scammer Bingo: In Editor";
@@ -60,6 +66,26 @@ public class GetOS : MonoBehaviour
                 { "potions", platform },
                 { "userid", id },
                 { "memoryused", process.PrivateMemorySize64 }
+            });
+        }
+        if (!Application.isEditor)
+        {
+            Process process = Process.GetCurrentProcess();
+            Analytics.CustomEvent("UserInfo", new Dictionary<string, object>
+            {
+                { "potions", platform },
+                { "userid", id.ToString() },
+                { "memoryused", process.PrivateMemorySize64.ToString()}
+            });
+        }
+        if (Application.isEditor)
+        {
+            Process process = Process.GetCurrentProcess();
+            Analytics.CustomEvent("UserInfoEditorTest", new Dictionary<string, object>
+            {
+                { "potions", platform },
+                { "userid", id.ToString() },
+                { "memoryused", process.PrivateMemorySize64.ToString()}
             });
         }
     }